Vendor go-gibberish code
  - The original [go-gibberish](https://github.com/AlessandroPomponio/go-gibberish) dependency is in public archive
    state and unmaintained.
  - In order to be able to use this library as a dependency for projects
    like otel-collector-contrib and opentelemetry-ebpf-instrumentation,
    we need to ensure dependencies are properly maintained and
    up-to-date.
  - The vendored code will now be maintained as part of this library.
