+++ LICENSE
@@ -199,3 +199,10 @@
    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
    See the License for the specific language governing permissions and
    limitations under the License.
+
+-------------------------------------------------------------------------------
+
+This project also includes code under the MIT License:
+
+Portions of this software are © 2015 Rob Renaud, licensed under the MIT License.
+See LICENSES/MIT.txt for the full license text.
