+++ pkg/consts/accepted_characters.go
@@ -1,0 +1,16 @@
+// Originally from go-gibberish
+// Copyright (c) 2015 Rob Renaud
+// Licensed under the MIT License. See LICENSES/MIT.txt.
+//
+// Modifications copyright (c) 2025 Grafana Labs
+// Licensed under the Apache License, Version 2.0.
+
+// Package consts contains constants.
+package consts
+
+const (
+
+	// AcceptedCharacters is a string with all the letters
+	// in the English alphabet, plus a space.
+	AcceptedCharacters = "abcdefghijklmnopqrstuvwxyz "
+)
